Fixed action map handling, changed attack system
The fight button now works, the damage is not timed for the moment. The attack system has been transferred from the damage marker to the player itself to make things clearer.
Input handling is now functional, it had a bug where when the game started, all action maps were activated.