Added main menu and fixed Items menu
The main menu has a play and quit button, as well as animation and custom sprites. Transitions between scenes are to be made.
Fixed a bug where you could select the Check button while in the Item menu, and added a custom text when no food is left using the new text system